Consolidated encode/decode into fec.h
